refactor(core): 恢复质量基线并收敛结构
