+++ src/commands/executor.rs
@@ -393,11 +393,7 @@ fn parse_cell_reference(input: &str) -> Option<(usize, usize)> {
 
 #[cfg(test)]
 mod tests {
-    use std::path::PathBuf;
-
     use super::parse_cell_reference;
-    use crate::app::{AppState, InputMode};
-    use crate::excel::{Cell, Sheet, Workbook};
 
     #[test]
     fn parses_valid_cell_references() {
@@ -409,27 +405,5 @@ mod tests {
     fn ignores_commands_with_non_ascii_arguments() {
         assert_eq!(parse_cell_reference("addsheet 测试1"), None);
         assert_eq!(parse_cell_reference("测试1"), None);
-    }
-
-    fn app_with_sheet() -> AppState<'static> {
-        let mut data = vec![vec![Cell::empty(); 3]; 3];
-        data[1][1] = Cell::new("Name".to_string(), false);
-        data[1][2] = Cell::new("Name".to_string(), false);
-        data[2][1] = Cell::new("Ada".to_string(), false);
-        data[2][2] = Cell::new("10".to_string(), false);
-
-        let sheet = Sheet {
-            name: "Data".to_string(),
-            data,
-            max_rows: 2,
-            max_cols: 2,
-            is_loaded: true,
-        };
-
-        AppState::new(
-            Workbook::from_sheets_for_test(vec![sheet]),
-            PathBuf::from("scores.xlsx"),
-        )
-        .unwrap()
     }
 }
