+++ src/ui/mod.rs
@@ -1,4 +1,5 @@
 mod handlers;
 mod render;
+mod theme;
 
 pub use crate::ui::render::run_app;
